requirements.txt
flask
flask-cors
supabase
python-dotenv
gunicorn
cloudinary
psycopg2-binary
